HACK: Use hardcoded fi:format selector
This is a hack / proof of concept which may be rendered unnecessary
by changes in post-2.2 releases of Evergreen.

The intent is to allow searching for "E-Books" and "Large Print
Books" by switching to a "format" search filter instead of an
"item_type" filter.

We are hardcoding the choices of format, and generating the <select>
using item_format_selector.tt2, which is a modified version of
coded_value_selector.tt2.

Signed-off-by: Jeff Godin <jgodin@tadl.org>