Display iframe widget for AVOD titles
First attempt at an AVOD widget, displayed when the user is logged
in. For SSL concerns, a custom proxy cgi is used. The proxy cgi
decides if it should display the widget or a static thumbnail.

In theory, this could be used for other content providers as well.

Add a "Resident patrons click here to view content" link for widgets
when not logged in.

Signed-off-by: Jeff Godin <jgodin@tadl.org>