Add dept location selector (replaces item type)
Add optional item_locations_selector (similar to item_type_selector)
which can be used in place of the item_type_selector in areas where
we would like to use shelving locations and not item type as a
search filter.

This is similar to features which exist in future versions of
Evergreen, and may be a simple intermediate step which is later
replaced.

Triggered by the physical_dept 'ys' at the moment, we have a
hardcoded list of groups of shelving locations. If triggered with an
unknown department, the item_locations_selector defaults to a single
empty entry for all locations.

Signed-off-by: Jeff Godin <jgodin@tadl.org>