Port changes required for staff hold placement
Port changes required for staff to place holds for patrons in the
staff client.

A number of changes were introduced between 2.2 and 2.5 in the
opac/parts/place_hold.tt2 template. Port the most important changes
in order for staff to reliably place holds in the staff client.

Post-reality commit: these changes have been in production for TADL
since at least 2014-01-02.

Signed-off-by: Jeff Godin <jgodin@tadl.org>